SDK Updater: Refactor, polish and complete the progress dialog.
This splits the former ProgressTask in 2 parts:
ProgressDialog is just the SWT dialog and ProgressTask
is the task handling.

This helps avoiding confusing SWT Designer each time
I do a change in the logic. Plus it seems cleaner anyway.

The other thing this CL does is implement the pause/close
I originally wanted: the pause button is greyed once used
by the user whilst the task completes. After, it becomes
a "Close" button if there's a result text displayed.
Closing the window using the close box is now the same as
using the pause/close button (to avoid closing the dialog
with a running thread.)