Fix the DeviceChooserDialog table issues.
On Linux, the gtk table seems to resize itself automatically when the
layout is computed and forces the last column to resize itself at +18 pixels.
This has a problematic impact on this dialog as it is not resizable and is
made to match the size of the content. As the dialog is used time after time
the last column grows by 18 pixels at every use up to a point where it
doesn't fit the screen.
Since storing the columns size is not that useful and I couldn't find a
way to ignore this first resize, I'm just removing the width storage.