Add method profiling support to DDMS.
New button allows to start/stop tracing. When clicking stop, DDMS downloads
the trace file and starts Traceview.

Also refactored some common parts of the HPROF and tracing handlers into a
common class. The goal is to have a default, extensible implementation
of the HPROF handler that DDMS and the plug-in can reuse. This will reduce
duplicated code.

Change-Id: Ifc48926c7f6f1c3ea49a4aa94053664be83cbb06