Specify emma dump file location on the device explicitly
This is to avoid using default location for the emma dump file (coverage.ec)
which is generated on the device. Pointing out the custom location (which
for now is the same as the default one) is more straightforward and
will be important if the default location changes in the future.