Set optimization mode to 'debug' automatically for debuggable applications.
As a convenience to NDK users, many who are not used to debugging native code at all
and will find that gdb cannot set breakpoints properly or dump local variables by
default strange.

This also moves the computation of APP_MANIFEST and APP_DEBUGGABLE to add-application.mk
which is just more logical.

Change-Id: Ia801f3aeda6a87d91d92bfc618c2620afd1470aa