C++ system headers refresh.
The headers are moved out of the platforms directory, and into
$NDK/sources/cxx-stl/system. This will drastically ease using other
STL implementations like STLport or GNU Libstdc++ in the future.

libstdc++ is not a mandatory dynamic dependency for C-based modules,
instead, it is now added automatically to modules that contain
at least one C++ source.

Change-Id: I84a859fdd460f3e9849e07c667c101a67aa84263
Note: nothing here is specific to a given Android API level.