Add easy native debugging support through NDK_DEBUG=1
"ndk-build NDK_DEBUG=1" will force the build of a debuggable application.
This really copies gdbserver to the proper location, as if android:debuggable
was set to "true" in the app's manifest.

The value of NDK_DEBUG can be 0, 1, true or false, and will override the content
of the manifest. The main benefit from this change is that you don't have to
edit your manifest file just to rebuilt

Change-Id: I3fafb620189ac53a72b492c963832dd0c7f8b7d7