Fix cygwin build on Windows.
This introduces changes to the build system to deal
with the fact that our Windows toolchain binaries are
not Cygwin programs anymore. As a consequence, they
don't understand paths like /cygdrive/c/stuff.

Essentially, this forces path conversions whenever we're
sending a path to the toolchain binaries, through the new
'host-path' and 'host-c-includes' helper functions.

Also, it reformats the auto-generated dependency files
in order for them to be properly parsed by Cygwin's
GNU Make (see build/awk/convert-deps-to-cygwin.awk for
examples).

As a bonus, this change also simplifies the setup of
each toolchain by grouping common definitions under
build/tools/default-build-commands.mk

Change-Id: I5af99b63cb53b3fcb5e1008dfb764e1e934623e5