Add --incremental option to make-release.sh
The idea is that making a release from scratch is now very
long, so use the --incremental option in order to speed-up
incremental release builds.

This is only useful when debugging make-release.sh or any
of the scripts it uses. I.e. no need to re-download + re-patch
the sources if you modify rebuild-all-prebuilts.sh

Change-Id: I37dc5832a19f2a50eef4cc39103721172b68e2a7