Proper thread support for gdbserver.
WARNING: DEEP MAGIC AHEAD!

This patch allows the gdbserver binary that we build for the
NDK to properly support threads on Android 2.3 and higher.

For previous platforms, the binary will still be limited to
to only set breakpoints on the main thread (just like with
previous NDK releases).

The root cause of the problem is a platform bug that was only
fixed in Android 2.3, so there is little we can do in the NDK
to fix it (apart from hoping the fix is back-ported properly).

Now the hairy details:

- The patch provides sources for libthread_db.a which will
  get recompiled explicitely by build-gdbserver.sh.

  Note that libthread_db has no real stable ABI/API so we
  provide GDB version-specific implementations here
  (see sources/android/libthread_db/gdb-<version>)

- Both libthread_db.c versions have been modified to probe
  the system at runtime and detect if the platform bug is in
  effect.

  If it is, the library acts as if there are no threads in the
  process. The reason for this behaviour (i.e. instead of
  patching the gdbserver sources) is explained in comments
  inside libthread_db.c.

+ Local patches to make signal handling and libthread_db support work
  properly with gdbserver 6.6.

Change-Id: Idf3d71422d337f4384f51ca45cafe7f71fd8559d