download-toolchain-sources.sh: Add --git-date
The option allows you to specify a date for the toolchain
sources that are checked-out from android.git.kernel.org.

The default date is set in prebuilt-common.sh and allows us
to ensure that the patches we have under build/tools/toolchain-patches
do apply correctly.

The plan is to modify the default date if needed when updating the
toolchain patches.

Change-Id: I27b94390c19e385d2232ab97f0be193e8a5a4c0d