NDK tests: Add support for broken builds and runtime tests.
This is needed because not all our tests are complete yet,
but we want to keep them in the tree nonetheless.

This is achieved by testing for empty files in the test directory

BROKEN_BUILD means the test cannot be built yet.
BROKEN_RUN means the test cannot run sucesfully yet.

Change-Id: Iaa14260ac4522730aacbd44d5c87b698b5c92314