Support remote Darwin builds on Linux.
This patch adds a --darwin-ssh=<hostname> option to both
rebuild-all-prebuilt.sh and make-release.sh. This allows one
to launch a build of Darwin binaries through ssh from a Linux
machine.

The idea is to fully automate the rebuild of all toolchain
binary tarballs. When the "mingw32" package is installed, the
"make-release.sh" script should be able to generate binaries
for all three supported systems now (provided --darwin-ssh is
used).

Change-Id: I8a62b47d4431c231de38cb9a425833bf9e648c84

+ fix minor bugs in script options parsing.
+ fix minor bug in run-tests.sh