make-standalone-toolchain.sh: fix symlink copy bug.
This fixes a bug where installing a standalone toolchain for
a platform other than android-3 would result in dangling symlinks
being copied.

The fix itself copies the symlinked files to the install.

Also add a basic test to check that a standalone toolchain installation
works properly. Will add more in the future. You can call
tests/run-standalone-tests.sh to launch it for now.

Change-Id: Ic73e70fd2004a8d138ebb61fcbbcbfac207467b2