Fix GNU STL crash on Android 2.1 and older.
Due to a bug in the implementation of __cxa_finalize() that was only
fixed in Android 2.2, any code linked with gnustl_static would crash
when running on Android 2.1 or older.

This adds a local patch to libsupc++ to detect the problem at runtime
and avoid the crash. Note that this results in a leak of a thread-specific
storage key on these older platforms.

Change-Id: Ia40ea9dc3ea5b0d3d3c1c3fae0556778a16c8ef6