Fixed logic errors for generating system libraries
When the SYMBOL_EXCLUDES is set to match all characters the 'grep -v'
results in a non-zero exit code and the SYMBOLS_INCLUDE 'grep' is
never executed.
Example 'copy_system_shared_library liblog'

Added double-quotes around calls to 'extract_shared_library_functions'
so the variable assignments are set to the entire list of results, not
just the first data item in the list.

Added missing library 'copy_system_shared_library libjnigraphics' to
the API levle 8. Sample code 'bitmap-plasma' was failing to compile
due to missing symbols.

Change-Id: Ib6781e24446c93dcf9b54ca9748a0c1d68bc5de1
Signed-off-by: Mark D Horn <mark.d.horn@intel.com>