download-toolchain-sources.sh: Add --git-reference option.
This patch adds the --git-reference=<path> option to the script
used to download the toolchain sources. This is useful to speed
up the script's operation if you maintain a local copy of the
toolchain repositories on your own machine.

One can trivially create such a reference with the "repo" tool,
i.e.:

   cd toolchain-reference
   repo init -u git://android.git.kernel.org/toolchain/manifest.git
   repo sync

Now invoke the script with:

  /path/to/download-toolchain-sources.sh --git-reference=/path/to/toolchain-reference

Note that the reference is only used as a cache for recently
downloaded git objects, i.e. the script will always load any piece
not in the reference from the external git repository.

You can refresh the reference periodically with:

  cd toolchain-reference
  repo sync

The checkout/branch state of the reference doesn't matter, git clone
will only use its .git content to pick object files from their SHA1.

Change-Id: I062ab38e98d6520e251dacad947dadbb3c9211e6