build-platforms.sh: Add --arch option.
This adds the --arch=<list> option to replace --abi=<list> (which is
still supported, but deprecated). The reason is simply that the script
expects a list of architectures, not ABI names.

Also, make the script accept comma-separated lists as well as
space-separated ones, this means that now the following are equivalent:

  build-platforms.sh --abi='arm x86'
  build-platforms.sh --abi='arm,x86'
  build-platforms.sh --arch='arm x86'
  build-platforms.sh --arch=arm,x86

The last form is the preferred format, because it makes it easier to
invoke the script with an architecture list from other scripts.

Change-Id: Iad3ccf79470b221e3fa56173f55b8db66687d53b