dev-system-import.sh: Add --force option.
This adds a --force option to the dev-system-import.sh script
to force it to copy all built platform files to a single
android-$PLATFORM directory under development/ndk.

The idea is to use this as:

  dev-system/import.sh --force --arch=x86 9

In order to sanitize the headers and library files for x86
support.

Also, ensure that gdbserver is linked statically with the android-9
platform libraries for both arm and x86 by default now (the result is
a static executable that will run on previous releases of the platform).

Change-Id: I316d6d62a786145773a14a4762dc057531928b75