Fix build-stlport.sh
Remove the compiler warning about the size and type of __dso_handle which
couldn't be found. Turns out we didn't define CRTBEGIN_SO_O and CRTEND_SO_O
properly.

Fix the --package-dir option processing. The script was not using proper
paths when calling pack_archive/tar, which resulted in the file placed
at the wrong location inside the final archive.

+ Make build commands visible with --verbose --verbose
+ Remove double declaration of -j<number> option

Change-Id: I179bd101e047ae71b13c0dfc081d063342b60b49