make-standalone-toolchain.sh: Fix --arch=x86
The script mistakenly copied the ARM GNU libstdc++ binaries into
the x86 toolchain installation directory. The end result was that
one could not build C++ programs with it.

Document that a minor issue with the standalone arm toolchain linker
is fixed (regarding the __dso_handle symbol). The fix is really in
https://review.source.android.com/24619

+ Modify tests/run-tests.sh and tests/run-standalone-tests.sh to
  use /tmp/ndk-$USER/tests as their working directory.

Change-Id: I0067d8a05306e4f999b1155d2e2de6659b17fa58