Support project.properties file in project path.
This patch modifies the NDK build system to support a file
named 'project.properties' in the application's project directory
to hold the value of the target platform.

If the file is not present, 'default.properties' will be searched
instead.

This patch is in preparation of future changes to the SDK tools,
which will rename/split default.properties and build.properties.

Change-Id: I17da187f495a2dec21b1b8c7cbfd716ae7878f8b