Support APP_ABI=all in Application.mk
This patch allows the use of 'APP_ABI := all' in your
Application.mk or on the command-line to indicate that you
want to build for all NDK-supported ABIs.

This is shorter than typing the full list of known ABIs,
and allows for an easy upgrade path when we'll introduce
new supported ABIs in the future (if any).

Change-Id: Ib5515727323fe1e3498b89acd0200969e2f793c5