Use shorter source/object/binary paths when possible.
This patch modifies the way the NDK build script computes the
value of NDK_PROJECT_PATH when it is not already defined.

If we are in the project's top-level directory when ndk-build
is invoked, we can use the value '.' for NDK_PROJECT_PATH.

This reduces the length of all paths passed to build commands
(source files, object files, binaries), which can be useful
for modules with a ton of sources (especially on Cygwin where
the maximum command-line lenght is quite limited).

Change-Id: I35a60df243c156690dc3020983b6232cd9b92ec6