Simplify non-Cygwin source compiles.
We don't need to rename the auto-generated dependency file when
we're not running inside Cygwin, so simplify operations.

In other words, before this patch, we would compile a source file
on Unix with something like:

  gcc -c -MF foo.d.org   -o foo.o foo.c &&
  mv foo.d.org foo.d

Now we will simply do:

  gcc -c -MF foo.d -o foo.o foo.c

Compilation on Cygwin isn't changed.

This makes the build commands dumped with V=1 easier to read.

Change-Id: Ic6f256ecb779de6c75ff940a1f14680bfd87127b