builder-funcs.sh: new shell functions
This patch introduces build/tools/builder-funcs.sh which contains
several functions that can be used to build source files into
static or shared libraries either directly or with the help of
an auto-generated Makefile (in order to allow for parallel builds).

Most of the functions come originally from build-stlport.sh which
has been modified to use the new stuff in builder-funcs.sh

Other NDK scripts will use this in the future to rebuild other
prebuilt binaries.

+ Support for NDK_CCACHE to speed up rebuilds. E.g. build-stlport.sh
  now takes 0.9s on my machine (used to be 26s before that).

Change-Id: I8ef97e62987f6aac6713821455d9a1ddcbbde796