Add NDK_CCACHE support to GCC build
This patch allows host rebuild scripts to use NDK_CCACHE when
it is defined.

Unfortunately, this doesn't speed build-gcc.sh tremendously on
my machine, because most of the build time is done with the second-phase
"xgcc" binary that is generated during the first phase. Still shaves off
20s out of 4m40s on my machine with -j32.

Improvements should definitely be better for other host binaries.

Change-Id: I7c0c8cfcab9600dc3317f1c9569602a457b88d57