Define HOST_OS as 'cygwin' on Cygwin
This patch changes the definition of HOST_OS to 'cygwin' when
we detect that the build system is invoked from a Cygwin-compatible
GNU Make program.

The old value of 'windows' will now be used for the case where the
build system is invoked from a vanilla Win32 GNU Make binary, e.g.
from a standard Windows shell.

First step towards getting rid of Cygwin.

Change-Id: I4f355a5d254b8d0f544ab6de60c0d7ff85ca23ec