prebuilt-common.sh: allow inclusion from any script
This patch allows any script to source prebuilt-common.sh
properly. Before that, only patches in the same directory,
i.e. build/tools/, could source it without errors.

Note: To do that, the parent script must define NDK_BUILDTOOLS_PATH
to point to the build/tools/ directory. If the variable is not defined,
prebuilt-common.sh assumes that it is sources from build/tools/

Change-Id: If91360a0a10571f968049003c8f7192ad98c7f5c