build-host-sed.sh: new script to build sed from sources
This patch adds build-host-sed.sh that can be used to rebuild
host versions of the sed tool, which will be placed under
prebuilt/<system>/bin/

This is mainly for native Win32 support, since we can't rely
on having a working sed in this environment.

A future patch will modify the NDK build system to use the
prebuilt binary automatically.

+ Add prebuilt/ to .gitignore
+ Remove prebuilt/ in dev-cleanup.sh

Change-Id: Ibe3561c189e90dde68023b7cbe1e9dc03dff1004