gnustl_shared: New C++ runtime
This patch fixes build-gnu-libstdc++.sh to work properly
and generate libgnustl_shared.so from the GNU libstdc++ sources.

This is done by a custom patch to the gcc toolchain source tree
to force the generation of libgnustl_shared.so from the same objects
than libstdc++.so. This is required because there is no way to change
the internal DT_SONAME of a given shared library, so a simple rename
will not work.

We also ensure that libgnustl_static.a is properly generated and
used by the make-standalone-toolchain.sh script (which uses
copy-libstdcxx.sh)

+ Update definitions.mk to make "gnustl_shared" available in APP_STL
+ Update documentation to reflect that STLport now supports RTTI
+ Update documentation to reflect the new gnustl_shared runtime

Change-Id: Ie3ffaf825ba2d65eca027ec0be5e320d0ee39cbc