ndk-gdb: Support activities with multiple intent categories
This is a fix for http://code.google.com/p/android/issues/detail?id=9400

The main intent-filter in an activity can have multiple categories.
The previous awk script (extract-launchable.awk) only used the last
category to determine whether an activity was launchable.

This modifies it to accept any activity with a main intent-filter
that has at least one category of LAUNCHABLE.

Change-Id: Ia9a6898470acd4cd067d8f6edea8093595f1191c