Minor native win32 fixes
This fixes a couple of issues in the native win32 build
support. Mainly reduce verbosity of host cp/rm commands
and deals with cmd.exe program's limitations.

We really need our own implementations of "rm" and "mkdir"
in the toolbox to avoid this kind of hack, but for now this
is good enough.

Change-Id: I07b4ad977357de7eeae5217b36c7854765fc0f42