Update 'exceptions-crash' unit test.
Update the test so that it actually crashes on Cupcake and Eclair.
This unit test is to track the bug described at

   http://code.google.com/p/android/issues/detail?id=20176

Note that there is no solution yet. It looks like this reveals
a bug in the system's dynamic linker that was only fixed in
Android 2.2 (probably related to weak symbols).

Change-Id: Ie8595016da6cd6fcba71a0d4b73cb0dedbf90a5a