Fix ndk-build arm/thumb target instruction selection.
Due to a small bug, ndk-build would fail to properly select
the thumb or ARM target instruction set when building a file
(either with or without LOCAL_ARM_MODE).

The problem occured in the following circumstances:

  - your project has two modules, which use different
    thumb/arm modes

  - each of these projects try to build the same source
    file, but in different modes.

Due to a bug, the mode selection of the first module was
applied when compiling the second one.

+ Added  unit test to check that everything works correctly
  for all ABIs and compilation modes (including NEON)

Change-Id: Ibcb9ac95b4b06e4c3f72e0a987926cf02f165bb9