ndk-build.cmd: Fix infinite loop / stack overflow
This patch fixes a problem which occured when calling the script
that was not the top-level project path (e.g. any sub-directory of it).

Due to the way GNU Make implements the $(dir ...) function on Windows,
the search in build-all.mk resulted in an infinite recursive loop
which ended with a stack overflow after some time.

Change-Id: I9c863bfd5c6aaeb87aa7acbf46301e6f9ee9082d