test-stlport: enforce STLPORT_FORCE_REBUILD
This patch ensures that we rebuild the STLport library from sources
whenever we want to run the full STLport test suite. This is useful
to ensure that any local change we make to the sources is reflected
when running the test (i.e. we don't want to use the prebuilt
libraries).

Note that a few tests in the suite are still currently broken.

Change-Id: If9cfc8b8fcd3aadbf4922a2811b424803fb2a5f1