ndk-gdb: Fix debugging of binaries generated with NDK_DEBUG=1
This patch ensures that ndk-gdb will always be able to find the
gdbsetup.ini file that was generated when calling with ndk-build NDK_DEBUG=1

In other words, when you did the following:

    $NDK/ndk-build NDK_DEBUG=1   # generate debug native machine code
    ant debug install            # install debug .apk to device
    $NDK/ndk-gdb --start

The operation would fail because ndk-gdb didn't find the location of
the gdbsetup.ini that was copied to $PROJECT/libs/$ABI/ by ndk-build.

This patch fixes it.

Change-Id: Iebeb4351b6ef6e6efa9ccad95ef7fe20a23b77d5