Fix standalone toolchain C++ linking issue.
Gets rid of a regression introduced with NDK r7

This patches gets rid of libstdc++.so in the standalone
toolchain installation, because when used at link time,
it generated executables and shared libraries that could
not be properly loaded by the dynamic linker.

Instead, one has to link with -lgnustl_shared now to be
able to use the shared version of the GNU STL with the
standalone toolchain. You can still use -lstdc++ but this
will always use the static version of the library, just
as before NDK r7.

+ fix tests/run-standalone-tests.sh

Change-Id: Ib992ce6f5a97bbd02cc318748636dca766b80857