build-host-make.sh: Copy sources to temp directory before build.
This patch modifies the build-host-make.sh to force it to copy the
GNU Make source directory to a temporary location before the build.

This is because the GNU Make makefiles will modify the source
directory (to apparently update the documentation). As a consequence,
each build did modified a few files un-necessarily, requiring a git
checkout to cleanup.

Change-Id: I5fbca5cab6d3c73f7422e98063b56fc8965f29f3