Do not force exceptions and rtti when using GNU libstdc++
This patch changes the default exported flags for both
gnustl_static and gnustl_shared. Previously, they were always
set to -fexceptions and -frtti, but this resulted in a few
problems.

Now, the flags are empty by default, but you can use
APP_GNUSTL_FORCE_CPP_FEATURES in your Application.mk to
revert to the old behaviour, in case you need it.

+ Add four unit tests to check that the build system works
  correctly with regards to this new variable definition

+ Documentation updates.

+ Modify tests/run-tests.sh to recognize a 'BUILD_SHOULD_FAIL'
  tag in a unit test directory. If such a file is present,
  the build _should_ fail for the test to pass.

  This is required because there is no way to reliably detect
  at compile time that -frtti is _not_ used, so we actually
  require the opposite from our unit test sources.
  (see gnustl-force-none and gnustl-force-exceptions).

Change-Id: Ia6fd236284c942bc0cb8fc0b49d9344ab68fe7ee