Remove hard-coded ABI names from build scripts.
This patch removes some of the hard-coded ABI names from our build
scripts. Now, each toolchain's config.mk is expected to provide a
TOOLCHAIN_ARCH definition corresponding to the target architecture
it supports.

This makes it easier to add support for new architectures or ABIs
in the NDK, just write a config.mk for your toolchain like:

   TOOLCHAIN_ARCH := myarch
   TOOLCHAIN_ABIS := myabi [myabi2]

Change-Id: Ib2f89102b343390fb16f66f63ee7e26f852701c2