build-gcc.sh: Fix Windows canadian-cross build for binutils 2.21
For some exotic reason, binutils 2.21 will fail to build properly
in canadian-cross configuration (i.e. generating ARM-targetting
Windows toolchain binaries on Linux)

Experimentation showed that this was due to the fact that CC and
CXX were defined in the environment at the wrong time. To solve this
we need to unset these variables and instead rely on a host-specific
toolchain named 'i586-pc-mingw32msvc-gcc'.

Note that this is different from 'i586-mingw32msvc-gcc' which is
the name of the 'mingw32' toolchain installed on Debian/Ubuntu.

To solve this problem, and still be able to build binutils 2.19,
we generate a wrapper toolchain with the right name.

This also means we can't use ccache when using --mingw to build
the intermediate host cross-compiler during the canadian build
(i.e. the one that is used to generate the ARM binaries like
libgcc.a during the build).

Change-Id: Ia75b6b651762e5d262635c6832ef4df827539fb2