stlport: Use __ANDROID__ to test for Android platform
This patch changes a few lines of STLport to ensure that the
library uses the __ANDROID__ macro definition to check for the
Android platform, instead of ANDROID.

That's because __ANDROID__ is always defined by the toolchain
compiler, while ANDROID must be defined on the command-line
explicitely.

The NDK build system currently adds -DANDROID to all compiler
commands, but it's better not to rely on this.

Change-Id: Id0e643732be1aa8cd8881f10a5e946607e1b3ee7