check-armeabi-v7a-prebuilts: loosen checks due to binutils-2.19 bugs
It looks like the check-armeabi-v7a-prebuilts script cannot work
properly when looking at machine code generated with GCC 4.4.3 + binutils 2.19

Document the madness in the script, and make it pass with the binaries
that we're going to provide in the next release (even though they contain
invalid ELF aeabi tags). I have checked manually, by looking at the
disassembled output that they're good.

Once we switch to binutils 2.21 or better, we'll be able to make the script
work as expected and not worry about this problem anymore.

The reason why this wasn't detected earlier is because I was using a
GCC 4.4.3 + binutils 2.21 toolchain to generate my prebuilts.

I had to run the NDK rebuild + packaging scripts from scratch, and run
the test suite on the results to see the problem.

Change-Id: I1ac331fe2a9a6fb1a9c32b52b6c5b42afd18296b