Fixed for /bin/dash
The default shell changed to dash after Ubuntu 6.10 (*).
The following line #375 containing "$((" isn't a bashism but it
confuses dash into searching for matching "))"

  LIBS=$((cd $SYMDIR && 2>/dev/null ls *.functions.txt)

The error message (points to the wrong offending line) reads:

  ./build/tools/gen-platforms.sh: 385: Syntax error: Missing '))'

Insert a space to fix the problem.

(*) NOTE: The problem doesn't manifest itself in Gubuntu which links
    /bin/sh to bash

Change-Id: Ic3f0c445d3220440979c47898fa9a11688828aba