build-mingw64-toolchain.sh: new script
This patch adds a new script that is capable of
generating a new Linux-hosted toolchain that generates
Win32 and Win64 executables.

The patch takes cares of downloading all required sources
from official sites, configure them, and run the build.

This should allow us to provide 64-bit Windows toolchain
binaries in the future.

Change-Id: I6c28372f6ac26cf1a18618a21101bd216f3e2590
WARNING: The patch works on Linux. It fails on Darwin due to
         strange libtool breakage when building binutils, and
         has never been tested on other systems.