Update NDK local patchset
This patch modifies our local NDK patches by rebasing them against
a more recent version of the toolchain sources (2012-04-20, while
the old ones were based on 2011-02-23).

Note that this removes certain local patches that are now in the
toolchain tree.

This allows us to integrate many bugfixes into the toolchain build.

Note: This is the first in a series of several patches that will
      add support fro GCC 4.6 and GDB 7.1.x/7.3.x. For now, the
      local patches still only apply to gcc 4.4.3 and gdb 6.6.

If you want to check this patch, you should do the following:

  - download the source tree _before_ applying this patch:

      download-toolchain-sources.sh /tmp/ndk-src1

  - download the source tree _after_ applying this patch:

      download-toolchain-sources.sh /tmp/ndk-src2

  - compare the two source trees in details, e.g. with:

      meld /tmp/ndk-src1 /tmp/ndk-src2

All the differences are supposed to come from upstream.

Change-Id: I881b231bec3ac04d1cd1de298aa96e44b46e44bb