Port local NDK patches to GCC 4.6
This patch ports all local NDK patches to the GCC 4.6 codebase.
The changes should not affect the GCC 4.4.3 sources/patches.

After applying this, re-run download-toolchain-sources.sh, then
do the following to build binaries for GCC 4.6:

  build/tools/build-gcc.sh $SRC_DIR $NDK_DIR $TOOLCHAIN_NAME --binutils-version=2.21

Where $TOOLCHAIN_NAME can be:

  arm-linux-androideabi-4.6
  x86-4.6
  mipsel-linux-android-4.6

Note that specifying binutils 2.21 is mandatory (the build will fail with 2.19, which is the default).

+ Update download-toolchain-sources.sh to download the mpc sources necessary for 4.6

Change-Id: Id5ffb5ec5305c3e99d6cdcf26da66de3f6fe27fd