Add NDK_TOOLCHAIN_VERSION support.
This patch modifies the build system to recognize the
NDK_TOOLCHAIN_VERSION environment variable to override
the GCC toolchain version to use during a build.

This will be useful when we'll provide both GCC 4.4.3 and 4.6
binaries at the same time. GCC 4.6 will be the default, but
defining NDK_TOOLCHAIN_VERSION to "4.4.3" will allow you to
use the old GCC 4.4.3 one for all target architectures.

Note that NDK_TOOLCHAIN is still supported. Actually, if you
define it, NDK_TOOLCHAIN_VERSION will be ignored completely.

Change-Id: I762256bd16693d81c6642f9f0560fe67df6a4092