build-host-gcc.sh: New script to rebuild toolchain. (DO NOT SUBMIT)
This patch adds build-host-gcc.sh, a new script that can rebuild
our GCC-based toolchain from scratch, without using the AOSP
toolchain/build.git scripts.

NOTE: The script does *not* rebuild GDB at all (a future patch will
      provide build-host-gdb.sh to do so).

The script is capable of building *several* toolchains at once,
minimizing what needs to be done to speed up the whole process.
For example, here's how to rebuild the linux-x86 and windows
toolchains for both gcc 4.4.3 and gcc 4.4.6 on all of arm,x86 and mips:

  build/tools/build-host-gcc.sh --toolchain-src-dir=/path/to/toolchain/src \
    --systems=linux-x86,windows \
    arm-linux-androideabi-4.4.3 \
    arm-linux-androideabi-4.6 \
    x86-4.4.3 \
    x86-4.6 \
    mipsel-linux-android-4.4.3 \
    mipsel-linux-android-4.6

This operations takes 31 minutes on a 16-core 2.4 GHz z600, not bad
for something that generates 12 different toolchains.

Define NDK_CCACHE=ccache in your environment to speed-up rebuilds.
On my machine, the above command takes above

+ The script will build for the host system by default. This means
  linux-x86_64 on 64-bit Ubuntu 10.10, for example.

DO NOT SUBMIT yet. The script works well, but we need to polish
it a bit with a few missing features:

  - better --help documentation.
  - copying the license files to the installation directory.
  - supporting packaging (i.e. --package-dir=<path> option).

This script has been tested to build the following toolchains
succesfully:

  arm-linux-androideabi-4.4.3
  arm-linux-androideabi-4.6
  x86-4.4.3
  x86-4.6
  mipsel-linux-android-4.4.3
  mipsel-linux-android-4.6

For the following host systems:

  linux-x86
  linux-x86_64
  windows (see note below)
  darwin-x86
  darwin-x86_64

Regarding windows, the build succeeds with the i586-mingw32msvc-gcc
cross-toolchain installed on Linux.

I have tried to build with x86_64-w64-mingw32 in my path (the script
will pick it up automatically), but this fails horribly because the
libtool script thinks it's building on Cygwin, for some stupid reason.

Still investigating, we might have to update ltmain.sh in the toolchain
source tree to work-around this.

Change-Id: I00f5d16d8db729e7efdf0e138888f761936f3c4d