toolbox: Fix windows echo program
This fixes the echo program to only accept -n as a valid option.
Everything else is to be treated as part of the input string.

This fixes LOCAL_SHORT_COMMANDS on native Windows because often
echo is used to send compiler flags like -fpic and others to the
generated list file, which failed with the previous program.

Change-Id: I4d4886a490db6d833a5f145ee3d0b14b4116bbc7