Rewrite standalone toolchain tests
This patch rewrites the standalone toolchain tests completely
(the previous version didn't work properly anyway).

tests/standalone/run.sh can be used to check that a given
NDK toolchain works properly. It doesn't have to be a
standalone toolchain as long as you use the --sysroot=<path>
option.

Usage example:

  tests/standalone/run.sh --prefix=toolchains/arm-linux-androideabi-4.4.3/prebuilt/linux-x86/bin/arm-linux-androideabi-gcc

Note that the tests do not assume that a C++ STL is available
in the toolchain's installation directory. We will probably
add a few other tests to check these later.

Change-Id: Idac1cf8fda040773de771b29ca5b80587f0b900e