gen-platforms.sh: fix <android/api-level.h> generation.
This patch fixes a hideous bug in the function that auto-generates
the content of <android/api-level.h> for each platform/arch combo.

The issue is that before calling the generation function, the
file would be created as a symlink to the older version by default
(unless you use the --no-symlinks option).

E.g. the scripts worked like this:

  1/ generate platforms/android-3/arch-arm/usr/include/android/api-level.h
     normally, setting __ANDROID_API__ to 3 as expected.

  2/ generate platforms/android-4/arch-arm/usr/include/android/api-level.h
     as a symlink to the version under platforms/android-3/... !

  3/ generate the android-4 api-level.h. When writing to it, the content
     will go directly into the android-3 file that the symlinks points to.

In the end, the android-3 file gets an __ANDROID_API__ of 14, corresponding
to the last API level processed/generated by the script.

To solve this, we simply remove the symlink before writing the auto-generated
content. Doh.

Change-Id: I74dc97ea60516c3a5d42a8ab6585381bcb4b067b