Add a test to build and run a trivial static executable.
This is to check that static executable generation works properly,
e.g. that our crtbegin_static.o sources aren't broken for some
reason.

Note that this test fails with the x86-4.4.3 toolchain generated
with build-host-gcc.sh at the moment (not with the one built with
build-gcc.sh). The symptom is:

   $NDK/toolchains/x86-4.4.3/prebuilt/linux-x86/bin/i686-linux-android-strip --strip-unneeded  ./libs/x86/static_main
   BFD: BFD (GNU Binutils) 2.19 internal error, aborting at /tmp/ndk-digit/src/binutils/binutils-2.19/bfd/elf.c line 4655 in assign_file_positions_for_non_load_sections

   BFD: Please report this bug.
   make: *** [libs/x86/static_main] Error 1
   make: *** Deleting file `libs/x86/static_main'

So it looks like a binutils-2.19 crash when 'strip' is called.

Change-Id: I84ed6ba9fceee84d30ad28885bbf6145397e0144